pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "incident_triage"
version = "0.1.0"
description = "Incident Response Triage - OpenEnv RL Environment"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["openai"]
